server/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn>=0.21.1

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Blockchain integration
web3>=6.0.0

# Task queue
celery>=5.2.7
redis>=4.5.0

# Testing
pytest>=7.3.0
httpx>=0.24.0